# openapi.yaml
openapi: 3.0.0
paths:
  /users:
    get:
      operationId: getUserList
      summary: ''
      parameters: []
      responses:
        '200':
          description: ユーザー一覧取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPresenter'
      tags: &ref_0
        - user
      security: &ref_1
        - &ref_2
          bearer: []
    post:
      operationId: createUser
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ユーザー登録
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPresenter'
      tags: *ref_0
      security: *ref_1
  /users/{id}:
    get:
      operationId: getUser
      summary: ''
      parameters:
        - name: id
          required: true
          in: path
          description: userId
          schema:
            type: string
      responses:
        '200':
          description: ユーザー詳細取得
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPresenter'
      tags: *ref_0
      security:
        - *ref_2
        - bearer: []
    patch:
      operationId: updateUser
      summary: ''
      parameters:
        - name: id
          required: true
          in: path
          description: userId
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ユーザーの情報更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPresenter'
      tags: *ref_0
      security: *ref_1
    delete:
      operationId: UserController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: userId
          schema:
            type: string
      responses:
        '204':
          description: ユーザー削除
      tags: *ref_0
      security: *ref_1
  /recipes:
    get:
      operationId: getRecipe
      summary: ''
      parameters: []
      responses:
        '200':
          description: レシピ一覧取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePresenter'
      tags: &ref_3
        - recipes
      security: &ref_4
        - bearer: []
    post:
      operationId: bulkInsertRecipes
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateRecipeDto'
      responses:
        '201':
          description: レシピ一括登録
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePresenter'
      tags: *ref_3
      security: *ref_4
  /recipes/favorite:
    get:
      operationId: getRecipeByFavorite
      summary: ''
      parameters: []
      responses:
        '200':
          description: お気に入りレシピ一覧取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePresenter'
      tags: *ref_3
      security: *ref_4
  /recipes/{id}:
    get:
      operationId: getRecipe
      summary: ''
      parameters: []
      responses:
        '200':
          description: レシピ詳細取得
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePresenter'
      tags: *ref_3
      security: *ref_4
    patch:
      operationId: updateRecipe
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeDto'
      responses:
        '200':
          description: レシピ更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePresenter'
      tags: *ref_3
      security: *ref_4
  /recipes/meta-data:
    get:
      operationId: getRecipeMetaData
      summary: ''
      parameters:
        - name: recipeIds
          required: true
          in: query
          description: レシピID
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: レシピメタデータ取得
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePresenter'
      tags: *ref_3
      security: *ref_4
  /recipe-books:
    post:
      operationId: createRecipeBook
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeBookDto'
      responses:
        '201':
          description: 料理本作成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePresenter'
      tags: &ref_5
        - recipe-books
      security: &ref_6
        - bearer: []
  /recipe-books/{id}:
    put:
      operationId: updateRecipeBook
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeBookDto'
      responses:
        '200':
          description: 料理本更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePresenter'
      tags: *ref_5
      security: *ref_6
  /invitations/invitation:
    post:
      operationId: invitationSpace
      summary: ''
      parameters: []
      responses:
        '200':
          description: スペース招待
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationPresenter'
      tags: &ref_7
        - invitations
      security: &ref_8
        - bearer: []
  /invitations:
    put:
      operationId: validateInvitation
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateInvitationDto'
      responses:
        '200':
          description: スペース参加の検証
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationPresenter'
      tags: *ref_7
      security: *ref_8
  /menus:
    get:
      operationId: getMenu
      summary: ''
      parameters: []
      responses:
        '200':
          description: 献立ー一覧取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuPresenter'
      tags: &ref_9
        - menus
      security: &ref_10
        - bearer: []
    post:
      operationId: createMenu
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuDto'
      responses:
        '200':
          description: 献立の作成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuPresenter'
      tags: *ref_9
      security: *ref_10
  /menus/pending:
    get:
      operationId: getPendingMenu
      summary: ''
      parameters: []
      responses:
        '200':
          description: 提案メニュー一覧取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuPresenter'
      tags: *ref_9
      security: *ref_10
  /menus/{id}:
    put:
      operationId: updateMenu
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuDto'
      responses:
        '200':
          description: 献立の更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuPresenter'
      tags: *ref_9
      security: *ref_10
    delete:
      operationId: deleteMenu
      summary: ''
      parameters: []
      responses:
        '200':
          description: 献立の削除
      tags: *ref_9
      security: *ref_10
  /media/image:
    post:
      operationId: uploadImage
      summary: ''
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadFilesDto'
      responses:
        '200':
          description: 画像のアップロード
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaPresenter'
      tags:
        - media
info:
  title: reciperu api app
  description: API documents
  version: '1.0'
  contact: {}
tags:
  - name: '1.0'
    description: ''
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    UserPresenter:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        imageUrl:
          type: string
      required:
        - id
        - name
        - imageUrl
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
      required:
        - name
        - imageUrl
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
      required:
        - name
        - imageUrl
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        imageUrl:
          type: string
      required:
        - id
        - name
        - imageUrl
    RecipePresenter:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        thumbnailUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        memo:
          type: string
        appName:
          type: string
        faviconUrl:
          type: string
        user:
          $ref: '#/components/schemas/User'
      required:
        - id
        - title
        - thumbnailUrl
        - imageUrls
        - memo
        - appName
        - faviconUrl
        - user
    CreateRecipeDto:
      type: object
      properties:
        title:
          type: string
        thumbnailUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        memo:
          type: string
      required:
        - title
        - thumbnailUrl
        - imageUrls
        - memo
    UpdateRecipeDto:
      type: object
      properties:
        title:
          type: string
        thumbnailUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        memo:
          type: string
      required:
        - title
        - thumbnailUrl
        - imageUrls
        - memo
    CreateRecipeBookDto:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateRecipeBookDto:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    InvitationPresenter:
      type: object
      properties:
        id:
          type: string
        token:
          type: string
        spaceId:
          type: string
      required:
        - id
        - token
        - spaceId
    ValidateInvitationDto:
      type: object
      properties:
        token:
          type: string
        spaceId:
          type: string
      required:
        - token
        - spaceId
    MenuPresenter:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        thumbnailUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        memo:
          type: string
        user:
          $ref: '#/components/schemas/User'
        createdAt:
          format: date-time
          type: string
      required:
        - id
        - title
        - thumbnailUrl
        - imageUrls
        - memo
        - user
        - createdAt
    CreateMenuDto:
      type: object
      properties:
        recipeId:
          type: number
        scheduledAt:
          format: date-time
          type: string
      required:
        - recipeId
        - scheduledAt
    UpdateMenuDto:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    UploadFilesDto:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
      required:
        - files
    MediaPresenter:
      type: object
      properties:
        url:
          type: array
          items:
            type: string
      required:
        - url